requirements.txt
fastapi
uvicorn
httpx
ruff
alembic
pydantic-settings
python-dotenv
psycopg2-binary
sqlmodel
asyncpg
pytest
pytest-asyncio
pytest-cov